pyproject.toml
[project]
name = "py-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.27.0",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
]
